Work for both sso and non-sso environments.

+++ app/controllers/application_controller.rb
@@ -25,7 +25,7 @@ class ApplicationController < ActionController::Base
   def current_user
     return @current_user if defined? @current_user
     @current_user = current_user_session && current_user_session.record
-    if @current_user && ( @current_user.net_id != sso_net_id )
+    if @current_user && sso_net_id && ( @current_user.net_id != sso_net_id )
       current_user_session.destroy
       @current_user_session = nil
       @current_user = nil
@@ -54,7 +54,7 @@ class ApplicationController < ActionController::Base
 
   def store_location
     return if self == UserSessionsController
-    session[:return_to] = request.request_uri
+    session[:return_to] = request.fullpath
   end
 
   def redirect_back_or_default(default)
